# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "test swagger with Operation api",
    "version": "1.0.0",
    "title": "test swagger"
  },
  "host": "localhost:3000",
  "schemes": [
    "http"
  ],
  "tags": [
    {
      "name": "test1",
      "description": "test swagger"
    },
    {
      "name": "multi",
      "description": "result A * B"
    },
    {
      "name": "divide",
      "description": "result A / B"
    },
    {
      "name": "plus",
      "description": "result A + B"
    },
    {
      "name": "minus",
      "description": "result A - B"
    }
  ],
  "paths": {
    "/test": {
      "get": {
        "tags": [
          "test1"
        ],
        "summary": "test get",
        "description": "testtest",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/test"
            }
          }
        }
      }
    },
    "/eig/{numberA}/{numberB}": {
      "get": {
        "tags": [
          "divide"
        ],
        "summary": "return A / B",
        "description": "intput A , B",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "numberA",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "numberB",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation and return result"
          },
          "400": {
            "description": "Invalid number value"
          }
        }
      }
    },
    "/multiAB": {
      "post": {
        "tags": [
          "multi"
        ],
        "summary": "return A * B",
        "description": "intput A , B",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/numberAB"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation and return result",
            "schema": {
              "$ref": "#/definitions/result"
            }
          },
          "400": {
            "description": "Invalid number value"
          }
        }
      }
    },
    "/FarMinus/{numberA}/{numberB}": {
      "get": {
        "tags": [
          "minus"
        ],
        "summary": "return A - B",
        "description": "intput A , B",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "numberA",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "numberB",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid number value"
          }
        }
      }
    },
    "/plusAB": {
      "post": {
        "tags": [
          "plus"
        ],
        "summary": "return A + B",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/numberAB"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/result"
            }
          },
          "400": {
            "description": "Invalid number value"
          }
        }
      }
    }
  },
  "definitions": {
    "test": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "ผ่านแล้ว"
        },
        "result": {
          "type": "string",
          "example": "Hello world Sat Dec 21 2019 03:11:22 GMT+0000 (Coordinated Universal Time)"
        }
      }
    },
    "numberAB": {
      "type": "object",
      "properties": {
        "A": {
          "type": "integer",
          "format": "int64",
          "example": "0"
        },
        "B": {
          "type": "integer",
          "format": "int64",
          "example": "0"
        }
      }
    },
    "result": {
      "type": "object",
      "properties": {
        "Math": {
          "type": "string",
          "example": "A operation B"
        },
        "result": {
          "type": "string",
          "example": "result"
        }
      }
    }
  }
}